正在实现管理员查询某天的时间线[已经完成]

+++ src/main/java/com/util/OnOffLineUtil.java
@@ -12,6 +12,6 @@ import lombok.Data;
 @Data
 public class OnOffLineUtil {
 
-    private long onLine;
-    private long offLine;
+    private int onLine;
+    private int offLine;
 }
